ENH: first pass of adding PF1K0_PM class
