+++ pcdsdevices/pneumatic.py
@@ -6,11 +6,13 @@ This Module contains all the classes relating to Pneumatic Actuators
 
 from lightpath import LightpathState
 from ophyd import Component as Cpt
+from ophyd import Signal
 from ophyd.status import Status
 
 from pcdsdevices.interface import BaseInterface, LightpathMixin
 
 from .analog_signals import FDQ
+from .inout import InOutPositioner
 from .signal import PytmcSignal
 
 
@@ -105,3 +107,40 @@ class BeckhoffPneumaticFDQ(BeckhoffPneumatic):
     """
     flow_meter = Cpt(FDQ, '', kind='normal',
                      doc='Device that measures PCW Flow Rate.')
+
+
+class PneumaticActuator(InOutPositioner):
+    states_list = ['RETRACTED', 'INSERTED', 'MOVING', 'INVALID']
+    in_states = ['INSERTED']
+    out_states = ['RETRACTED']
+    _invalid_states = ['MOVING', 'INVALID']
+    _unknown = False
+
+    state = Cpt(PytmcSignal, ':POS_STATE', io='i', kind='hinted')
+
+    in_sw = Cpt(PytmcSignal, ':IN', io='i', kind='normal')
+    out_sw = Cpt(PytmcSignal, ':OUT', io='i', kind='normal')
+    error = Cpt(PytmcSignal, ':ERROR', io='i', kind='normal')
+
+    in_cmd = Cpt(PytmcSignal, ':IN_CMD', io='io', kind='config')
+    out_cmd = Cpt(PytmcSignal, ':OUT_CMD', io='io', kind='config')
+    filter_type = Cpt(Signal, value='Unknown filter', kind='config')
+
+    done = Cpt(PytmcSignal, ':MOT_DONE', io='i', kind='omitted')
+
+    def __init__(self, prefix, *, name, transmission=1,
+                 filter_type='Unknown filter', **kwargs):
+        self._transmission = {'INSERTED': transmission}
+        super().__init__(prefix, name=name, **kwargs)
+        self.filter_type.put(filter_type)
+
+    def _do_move(self, state):
+        """
+        Override state move because we can't use the default.
+
+        Here we need to put 1 to the proper command pv.
+        """
+        if state.name == 'INSERTED':
+            self.in_cmd.put(1)
+        elif state.name == 'RETRACTED':
+            self.out_cmd.put(1)
